moved grey line to header and cleaned up a bit and deleted article file
