+++ code/src/index.css
@@ -36,6 +36,7 @@ a {
 
 .artist-name {
   color: #a0a0a0;
+  font-size: 14px;
 }
 
 .album-name a {
